Fixed pointer issue created while norminnetting
