+++ Makefile
@@ -5,7 +5,7 @@ MLX = ./incl/mlx/libmlx.a
 INC = -I ./incl
 OBJ_DIR = obj/
 CC = cc
-CFLAGS = -O3 -pthread -g $(INC) -Wall -Wextra -Werror
+CFLAGS = -O3 -pthread -g $(INC) -Wall -Wextra -Werror -fsanitize=address,undefined
 MFLAG = -L /usr/X11R6/lib -lXext -lX11 -lm
 RM = rm -f
 
@@ -34,7 +34,9 @@ SRC = srcs/parser/checks.c \
       srcs/utils_color.c \
       srcs/utils_vec.c \
       srcs/utils.c \
-      srcs/main.c \
+      srcs/utils_print.c \
+      srcs/utils_selection.c \
+      srcs/main.c
 
 OBJ = $(addprefix $(OBJ_DIR), $(SRC:.c=.o))
